When Natural Laziness meets Artifical Intelligence, or How to Build Agentic Systems/google-adk-demo/requirements.txt
google-adk
yfinance
psutil
